+++ packages/category.go
@@ -5,7 +5,9 @@
 package packages
 
 type Category struct {
-	Id    string `yaml:"id" json:"id"`
-	Title string `yaml:"title" json:"title"`
-	Count int    `yaml:"count" json:"count"`
+	Id          string `yaml:"id" json:"id"`
+	Title       string `yaml:"title" json:"title"`
+	Count       int    `yaml:"count" json:"count"`
+	ParentId    string `yaml:"parent_id,omitempty" json:"parent_id,omitempty"`
+	ParentTitle string `yaml:"parent_title,omitempty" json:"parent_title,omitempty"`
 }