+++ CHANGELOG.md
@@ -8,12 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking changes
 
+* Updated titles of some categories. [#914](https://github.com/elastic/package-registry/pull/914)
+
 ### Bugfixes
 
 ### Added
 
 * Forward requests from package-storage instead of doing http redirects. [#915](https://github.com/elastic/package-registry/pull/915)
 * Update default value for `proxy-url` parameter to be Elastic Package Registry production. [#899](https://github.com/elastic/package-registry/pull/899)
+* Add additional categories and subcategories. [#914](https://github.com/elastic/package-registry/pull/914)
+* Support subcategories. Include parent category in categories API. [#914](https://github.com/elastic/package-registry/pull/914)
 * Update Go runtime to 1.19.3. [#919](https://github.com/elastic/package-registry/pull/919)
 
 ### Deprecated